requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
-e git+https://github.com/ArcadeAI/arcade-ai.git@c980235535e275d1d1b59c0fb8021cdc4b9bc26a#egg=arcade_ai&subdirectory=arcade
-e git+https://github.com/ArcadeAI/arcade-ai.git@c980235535e275d1d1b59c0fb8021cdc4b9bc26a#egg=arcade_google&subdirectory=toolkits/google
arcadepy==0.2.2
beautifulsoup4==4.12.3
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
distlib==0.3.9
distro==1.9.0
dulwich==0.21.7
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.5
jiter==0.8.2
keyring==24.3.1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
oauthlib==3.2.2
openai==1.58.1
packaging==24.2
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
proto-plus==1.25.0
protobuf==5.29.2
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
pyproject_hooks==1.2.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.11.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
toml==0.10.2
tomlkit==0.12.5
tqdm==4.67.1
trove-classifiers==2024.10.21.16
typer==0.9.4
types-python-dateutil==2.9.0.20241003
types-pytz==2024.2.0.20241003
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.0
xattr==1.1.0
